+++ CHANGELOG.md
@@ -15,8 +15,23 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Updates
 
 - [Updates `Template` component(s)](https://newyorkpubliclibrary.atlassian.net/browse/DSD-1884) and styles per TAD.
+- Updates the values for the breakpoints and the associated docs.
+- Updates the `viewport` preview options in Storybook to align with the Reservoir breakpoints.
 
 ## Prerelease
+
+## 3.5.3 (January 30, 2025)
+
+### Adds
+
+- Adds the `useFormatNumber` hook.
+
+### Updates
+
+- Updates the `Pagination` component to handle 4 digit page counts.
+- Updates the `Heading` component to add `line-height` styles for mobile.
+- Updates the `Breadcrumbs` component to make link text screen readable on mobile breakpoints.
+- Updates the `MultiSelect` component to adjust the text and size styles for the search input field.
 
 ## 3.5.2 (January 16, 2025)